chore: implement sale entity

+++ app/feature/product/lib/presentation/widgets/product_detail_content.dart
@@ -135,7 +135,7 @@ class ProductDetailContent extends StatelessWidget {
                                 storeId: product.storeId,
                                 name: name,
                                 saleNet:
-                                    price.isEmpty ? null : int.parse(price),
+                                    price.isEmpty ? null : double.parse(price),
                               ),
                             ),
                           );
